feat: ft-wave1-petri-simple-run-004 - Consume mapped deletion-owned scenarios and close verification gates
