+++ internal/migrationledgercheck/check.go
@@ -99,7 +99,6 @@ var ExpectedDeletionOnlyBatches = []string{
 	"runtime_api-delete-09-workers-resilience",
 	"runtime_api-delete-10-observability-product",
 	"runtime_api-delete-11-wrong-layer",
-	"smoke-delete-01-orchestration-petri",
 	"smoke-delete-03-factory-definitions",
 	"smoke-delete-04-transport-cli",
 	"smoke-delete-05-factory-packaged",
@@ -119,7 +118,6 @@ var ExpectedDeletionOnlyBatches = []string{
 	"workflow-delete-09-orchestration-javascript",
 	"guards_batch-delete-01-work-relationships",
 	"guards_batch-delete-02-guards-global",
-	"guards_batch-delete-03-orchestration-dispatch",
 	"guards_batch-delete-04-resources-concurrency",
 	"guards_batch-delete-05-resources-fairness",
 	"guards_batch-delete-06-resilience-batch",
