--- tests/functional/runtime_api/api_ootb_experience_smoke_long_test.go
@@ -1,113 +1,0 @@
-//go:build functionallong
-
-package runtime_api
-
-import (
-	"testing"
-	"time"
-
-	"github.com/portpowered/infinite-you/internal/testutil"
-	"github.com/portpowered/infinite-you/pkg/services/work"
-	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
-	"github.com/portpowered/infinite-you/tests/functional/internal/support"
-)
-
-func TestOOTBExperience_APIPreseededSimplePipelineCompletes(t *testing.T) {
-	support.SkipLongFunctional(t, "slow OOTB API simple pipeline sweep")
-
-	dir := support.ScaffoldFactory(t, simplePipelineConfig())
-	testutil.WriteSeedRequest(t, dir, work.SubmitRequest{
-		WorkTypeID: "task",
-		TraceID:    "trace-ootb-001",
-		Payload:    []byte(`{"title":"Hello World"}`),
-	})
-
-	server := startFunctionalServer(t, dir, true)
-
-	initialStatus := getGeneratedJSON[factoryapi.StatusResponse](t, server.URL()+"/status")
-	if initialStatus.FactoryState == "" {
-		t.Fatal("GET /status returned an empty factory_state during first-run smoke")
-	}
-
-	token := waitForGeneratedWorkTypeComplete(t, server.URL(), "task", 10*time.Second)
-	if stringPointerValue(token.WorkTypeName) != "task" {
-		t.Fatalf("GET /work completed work type = %q, want task", stringPointerValue(token.WorkTypeName))
-	}
-	if generatedWorkStateName(token.State) != "complete" {
-		t.Fatalf("GET /work completed state = %#v, want complete", token.State)
-	}
-
-	status := getGeneratedJSON[factoryapi.StatusResponse](t, server.URL()+"/status")
-	if status.TotalTokens != 1 {
-		t.Fatalf("GET /status total_tokens = %d, want 1", status.TotalTokens)
-	}
-	if status.Categories.Terminal != 1 {
-		t.Fatalf("GET /status terminal count = %d, want 1", status.Categories.Terminal)
-	}
-	if status.Categories.Failed != 0 {
-		t.Fatalf("GET /status failed count = %d, want 0", status.Categories.Failed)
-	}
-}
-
-func TestOOTBExperience_APIPreseededTwoStagePipelineCompletes(t *testing.T) {
-	support.SkipLongFunctional(t, "slow OOTB API two-stage pipeline sweep")
-
-	dir := support.ScaffoldFactory(t, ootbTwoStagePipelineConfig())
-	testutil.WriteSeedRequest(t, dir, work.SubmitRequest{
-		WorkTypeID: "task",
-		TraceID:    "trace-ootb-multistage-001",
-		Payload:    []byte(`{"title":"Multi-stage test"}`),
-	})
-
-	server := startFunctionalServer(t, dir, true)
-
-	token := waitForGeneratedWorkTypeComplete(t, server.URL(), "task", 10*time.Second)
-	if stringPointerValue(token.WorkTypeName) != "task" {
-		t.Fatalf("GET /work completed work type = %q, want task", stringPointerValue(token.WorkTypeName))
-	}
-	if generatedWorkStateName(token.State) != "complete" {
-		t.Fatalf("GET /work completed state = %#v, want complete", token.State)
-	}
-
-	status := getGeneratedJSON[factoryapi.StatusResponse](t, server.URL()+"/status")
-	if status.TotalTokens != 1 {
-		t.Fatalf("GET /status total_tokens = %d, want 1", status.TotalTokens)
-	}
-	if status.Categories.Terminal != 1 {
-		t.Fatalf("GET /status terminal count = %d, want 1", status.Categories.Terminal)
-	}
-}
-
-func ootbTwoStagePipelineConfig() map[string]any {
-	return map[string]any{
-		"workTypes": []map[string]any{{
-			"name": "task",
-			"states": []map[string]string{
-				{"name": "init", "type": "INITIAL"},
-				{"name": "processing", "type": "PROCESSING"},
-				{"name": "complete", "type": "TERMINAL"},
-				{"name": "failed", "type": "FAILED"},
-			},
-		}},
-		"workers": []map[string]string{
-			{"name": "worker-a"},
-			{"name": "worker-b"},
-		},
-		"workstations": []map[string]any{
-			{
-				"name":      "step-one",
-				"worker":    "worker-a",
-				"inputs":    []map[string]string{{"workType": "task", "state": "init"}},
-				"outputs":   []map[string]string{{"workType": "task", "state": "processing"}},
-				"onFailure": []map[string]string{{"workType": "task", "state": "failed"}},
-			},
-			{
-				"name":      "step-two",
-				"worker":    "worker-b",
-				"inputs":    []map[string]string{{"workType": "task", "state": "processing"}},
-				"outputs":   []map[string]string{{"workType": "task", "state": "complete"}},
-				"onFailure": []map[string]string{{"workType": "task", "state": "failed"}},
-			},
-		},
-	}
-}
