+++ tests/functional/smoke/config_driven_execution_test.go
@@ -3,7 +3,6 @@ package smoke
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -163,45 +162,6 @@ func dereferenceGlobalConfigValue(value *string) string {
 		return ""
 	}
 	return *value
-}
-
-func TestConfigDrivenExecution_HappyPath(t *testing.T) {
-	dir := testutil.CopyFixtureDir(t, support.LegacyFixtureDir(t, "happy_path"))
-
-	testutil.WriteSeedFile(t, dir, "task", []byte(`{"title": "Config-driven happy path"}`))
-
-	provider := testutil.NewMockProvider(
-		workerexecution.InferenceResponse{Content: "Step one done. COMPLETE"},
-		workerexecution.InferenceResponse{Content: "Step two done. COMPLETE"},
-	)
-
-	status := runFactoryThroughCustomerProcess(t, dir, provider)
-	if status.Categories.Terminal != 1 || status.Categories.Failed != 0 {
-		t.Fatalf("status categories = %+v, want one terminal work item", status.Categories)
-	}
-
-	if provider.CallCount() != 2 {
-		t.Errorf("expected provider called 2 times, got %d", provider.CallCount())
-	}
-}
-
-func TestConfigDrivenExecution_HappyPathFailureRouting(t *testing.T) {
-	dir := testutil.CopyFixtureDir(t, support.LegacyFixtureDir(t, "happy_path"))
-
-	testutil.WriteSeedFile(t, dir, "task", []byte(`{"title": "Will fail"}`))
-
-	provider := testutil.NewMockProviderWithErrors(
-		[]workerexecution.InferenceResponse{{Content: ""}},
-		[]error{fmt.Errorf("something went wrong")},
-	)
-
-	status := runFactoryThroughCustomerProcess(t, dir, provider)
-	if status.Categories.Failed != 1 || status.Categories.Terminal != 0 {
-		t.Fatalf("status categories = %+v, want one failed work item", status.Categories)
-	}
-	if provider.CallCount() != 1 {
-		t.Errorf("expected failed provider called once, got %d", provider.CallCount())
-	}
 }
 
 func TestConfigDrivenExecution_AddWorkType(t *testing.T) {
